Fix tests after refactor
This completes work towards #803

- Fix import in test_class_command.py.
- Moved tests in test_analyze_common into a new file.
  test_analyze_default_filter.py that correspond to the filter
  funcitons.
- Renamed test_analyze_dockerfile to
  test_analyze_default_dockerfile_parse to reflect the tested
  functions.
- Renamed all the tests in ci/test_files_touched.py accordingly
